Update for v1_3
Moved codec configuration to software. In software write 0x34XXXX where
XXXX is the 16 bit register address and data to I2C data pio. Then write
1 to I2C start flag. Wait for I2C end flag to equal 0. Clear I2C start
flag and wait for I2C end flag to be cleared by hardware. Important
files are located in the Output Files folder. The register values the
codec needs to be set to are listed in clock_500.v