Fixed Hardware Files and added Software used to test
All the hardware output files should be working now except for the
version of 1_5 that does audio config in the hardware. It writes to the
FIFO correctly, but nothing is playing. I think the problem might be the
audio config not getting set, but I'm not sure if it's a hardware or
software problem. It's just a intermediate test step, so I'm not going
to waste any more time trying to debug it. I've also included the
software files I used to test. All that's left is to write the software
which I'll start doing now. It should take way less time than the
hardware took.