Audio In Data; Hardware/software for Auto-tune
The audio in data text files are in the Audio In Data folder as well as
the MATLAB code I was using to do some quick analysis. I've also
included the new hardware configuration. This one has three FIFOs in
hardware: one for audio in, one for the auto-tuned audio out, and one
for the instrumental track. Version 1_8 is the serial version and v1_9
is the LCD screen version. I've also included the software code I've
been using to test with so far. I'm having some troubles getting audio
in to play correctly.